requirements.txt
numpy==1.24.3
matplotlib==3.8.0
keras==2.13.1
segmentation-models==1.0.1
scikit-learn==1.3.1
opencv-python==4.8.1.78
laspy==2.5.1
tqdm==4.66.1
cvlib
fastapi
Pillow
openpyxl